Moved window manager and desktop environment class definitions to separate files
