+++ archey/constants.py
@@ -24,51 +24,6 @@ COLOR_DICT = {
     'clear': '\x1b[0m'
 }
 
-DE_DICT = {
-    'cinnamon': 'Cinnamon',
-    'gnome-session': 'GNOME',
-    'gnome-shell': 'GNOME',
-    'mate-session': 'MATE',
-    'ksmserver': 'KDE',
-    'xfce4-session': 'Xfce',
-    'fur-box-session': 'Fur Box',
-    'lxsession': 'LXDE',
-    'lxqt-session': 'LXQt'
-}
-
-WM_DICT = {
-    'awesome': 'Awesome',
-    'beryl': 'Beryl',
-    'blackbox': 'Blackbox',
-    'bspwm': 'bspwm',
-    'cinnamon': 'Cinnamon',
-    'compiz': 'Compiz',
-    'dwm': 'DWM',
-    'enlightenment': 'Enlightenment',
-    'herbstluftwm': 'herbstluftwm',
-    'fluxbox': 'Fluxbox',
-    'fvwm': 'FVWM',
-    'i3': 'i3',
-    'icewm': 'IceWM',
-    'kwin_x11': 'KWin',
-    'metacity': 'Metacity',
-    'musca': 'Musca',
-    'openbox': 'Openbox',
-    'pekwm': 'PekWM',
-    'qtile': 'QTile',
-    'ratpoison': 'ratpoison',
-    'scrotwm': 'ScrotWM',
-    'stumpwm': 'StumpWM',
-    'subtle': 'Subtle',
-    'monsterwm': 'MonsterWM',
-    'wingo': 'Wingo',
-    'wmaker': 'Window Maker',
-    'wmfs': 'Wmfs',
-    'wmii': 'wmii',
-    'xfwm4': 'Xfwm',
-    'xmonad': 'xmonad'
-}
-
 LOGOS_DICT = {
     Distributions.ARCH_LINUX: logos.ARCH_LINUX,
     Distributions.BUNSENLABS: logos.BUNSENLABS,
