--- archey/entries/__init__.py
@@ -1,15 +1,0 @@
-from .user import User
-from .hostname import Hostname
-from .model import Model
-from .kernel import Kernel
-from .uptime import Uptime
-from .shell import Shell
-from .terminal import Terminal
-from .packages import Packages
-from .temperature import Temperature
-from .cpu import CPU
-from .gpu import GPU
-from .ram import RAM
-from .disk import Disk
-from .lan_ip import LanIp
-from .wan_ip import WanIp
