+++ archey/entries/_constants.py
@@ -1,0 +1,44 @@
+DE_DICT = {
+    'cinnamon': 'Cinnamon',
+    'gnome-session': 'GNOME',
+    'gnome-shell': 'GNOME',
+    'mate-session': 'MATE',
+    'ksmserver': 'KDE',
+    'xfce4-session': 'Xfce',
+    'fur-box-session': 'Fur Box',
+    'lxsession': 'LXDE',
+    'lxqt-session': 'LXQt'
+}
+
+WM_DICT = {
+    'awesome': 'Awesome',
+    'beryl': 'Beryl',
+    'blackbox': 'Blackbox',
+    'bspwm': 'bspwm',
+    'cinnamon': 'Cinnamon',
+    'compiz': 'Compiz',
+    'dwm': 'DWM',
+    'enlightenment': 'Enlightenment',
+    'herbstluftwm': 'herbstluftwm',
+    'fluxbox': 'Fluxbox',
+    'fvwm': 'FVWM',
+    'i3': 'i3',
+    'icewm': 'IceWM',
+    'kwin_x11': 'KWin',
+    'metacity': 'Metacity',
+    'musca': 'Musca',
+    'openbox': 'Openbox',
+    'pekwm': 'PekWM',
+    'qtile': 'QTile',
+    'ratpoison': 'ratpoison',
+    'scrotwm': 'ScrotWM',
+    'stumpwm': 'StumpWM',
+    'subtle': 'Subtle',
+    'monsterwm': 'MonsterWM',
+    'wingo': 'Wingo',
+    'wmaker': 'Window Maker',
+    'wmfs': 'Wmfs',
+    'wmii': 'wmii',
+    'xfwm4': 'Xfwm',
+    'xmonad': 'xmonad'
+}
